added data feeding and added error handlings

+++ main.py
@@ -1,5 +1,6 @@
 import os
 import pandas as pd
+from pandas import ExcelWriter
 from flask import Flask, render_template, request, jsonify, send_file
 from flask_uploads import UploadSet, configure_uploads, DATA
 from werkzeug.utils import secure_filename
@@ -10,9 +11,9 @@ from time import sleep
 import functions as func
 import joblib
 from sklearn.feature_extraction.text import CountVectorizer
-from predict import load_model_predict
+# from predict import load_model_predict
 from train import train_model_save_joblib
-
+import xlsxwriter
 import numpy as np
 
 app = Flask(__name__)
@@ -63,28 +64,26 @@ def delete():
 
 @app.route('/merge', methods=['POST'])
 def merge():
+    output_file_path = None
     try:
         status = False
-
+        
         if not os.path.exists(directory_path):
             message = f"The system cannot find the path specified: {directory_path}"
             data_to_return = {'message': message, 'status': status}
             return jsonify(data_to_return)
-
+        
         bank_name = request.form['bank_name']
         
         merge_excel_folder = os.path.join(directory_path, "Merge-Excel")
-        # area_break_folder = os.path.join(directory_path, "Address")
+        
 
         if not os.path.exists(merge_excel_folder):
             os.makedirs(merge_excel_folder)
-
-        # if not os.path.exists(area_break_folder):
-        #     os.makedirs(area_break_folder)
         
         folder_path = os.path.join(directory_path, "Requests", bank_name)
         files = [f for f in os.listdir(folder_path) if os.path.isfile(os.path.join(folder_path, f)) and f.endswith('.xlsx')]
-
+        
         if not os.path.exists(folder_path):
             os.makedirs(folder_path)
 
@@ -100,7 +99,8 @@ def merge():
 
         mapping = {
             "REQUEST DATE": "DATE REQUESTED",
-            "REQUEST NAME": "REQUESTED BY"
+            "REQUEST NAME": "REQUESTED BY",
+            "ADDRESS TYPE": "ADD TYPE"
             # Add more mappings as needed
         }
 
@@ -108,20 +108,22 @@ def merge():
 
         total_files = len(files)
         work_progress = total_files + 6
+        work_progress = total_files + 6
 
         def set_progress(progress):
             sleep(0.75)
+            sleep(0.75)
             socketio.emit("update progress", progress)
-
+        
         set_progress(0)
 
+        # main.py
         for i, file in enumerate(files):
             excel_file_path = os.path.join(folder_path, file)
             index_header = func.get_index_of_header(excel_file_path, template_header)
-            work_book = pd.read_excel(excel_file_path, sheet_name=None, header=index_header)
-
-            for _, sheet_data in work_book.items():
-                for _, row in sheet_data.iterrows():
+            sheet_data = pd.read_excel(excel_file_path, sheet_name=0, header=index_header)
+            
+            for _, row in sheet_data.iterrows():
                     output_row = []
 
                     # Fill out necessary column from template
@@ -130,6 +132,7 @@ def merge():
                         for col_header, col_value in row.items():
                             if func.compare_string(header, col_header):
                                 value = col_value
+                                existing_headers.add(header)
                                 break
                         output_row.append(value)
 
@@ -138,6 +141,8 @@ def merge():
                     # Append other columns from the data frame
                     for col_header in row.keys():
                         mapped_header = func.map_header(col_header, mapping)
+
+                        # print(mapped_header, '-', mapped_header.lower() not in [h.replace('_', '').lower() for h in template_header], mapped_header not in existing_headers, not (col_header.strip() == "" or col_header.startswith("Unnamed")), func.compare_string(mapped_header, col_header))
 
                         if mapped_header.lower() not in [h.lower() for h in template_header] and mapped_header not in existing_headers and not (col_header.strip() == "" or col_header.startswith("Unnamed")) and func.compare_string(mapped_header, col_header):
                             datas[0].append(mapped_header)
@@ -150,12 +155,13 @@ def merge():
         output_work_book = pd.DataFrame(datas[1:], columns=datas[0])
         random_number = "".join([str(random.randint(0, 9)) for _ in range(4)])
         current_date = datetime.now().strftime("%Y-%m-%d")
-
+        
         output_file_name = f"Output-{bank_name}-{current_date}-{random_number}.xlsx"
         output_file_path = os.path.join(merge_excel_folder, output_file_name)
-
+        
         set_progress((total_files + 1) / work_progress * 100)
         
+        # Make addresses in 'ADDRESS' column uppercase
         # Make addresses in 'ADDRESS' column uppercase
         output_work_book['ADDRESS'] = output_work_book['ADDRESS'].str.upper()
         output_work_book.to_excel(output_file_path, index=False)
@@ -166,25 +172,44 @@ def merge():
         func.highlight_n_fill_missing_values(output_file_path, campaign_file_path)
 
         set_progress((total_files + 2) / work_progress * 100)
-        
+        print('LOADING MODEL')
         model = joblib.load('source\\model.joblib')
         output_work_book = pd.read_excel(output_file_path)
 
-        # Extract addresses from the 'ADDRESS' column
-        addresses = output_work_book['ADDRESS'].tolist()
+        set_progress((total_files + 3) / work_progress * 100)
+        print('PREDICTING')
+        # Define a mask for rows where 'AREA' and 'MUNICIPALITY' already have values
+        existing_mask = (output_work_book['AREA'].notna()) & (output_work_book['MUNICIPALITY'].notna())
+        # Extract non-empty addresses from the 'ADDRESS' column
+        addresses = output_work_book.loc[~existing_mask, 'ADDRESS'].tolist()
+        addresses = output_work_book.loc[~existing_mask & output_work_book['ADDRESS'].notna(), 'ADDRESS'].tolist()
+        # Create a new mask for rows that were used for prediction
+        prediction_mask = output_work_book['ADDRESS'].isin(addresses)
+        print(f'Number of addresses to predict: {len(addresses)}')  # Debugging line
 
         # Make predictions using the model
-        predictions = model.predict(addresses)
+        if addresses:
+            predictions = model.predict(addresses) 
+        else:
+            predictions = []
 
+        print(f'Number of predictions: {len(predictions)}')  # Debugging line
+
+        print('SPLITTING')
         # Split predictions into AREA and MUNICIPALITY
         area_munis = [prediction.split('-') for prediction in predictions]
 
-        # Update the DataFrame with predictions
-        output_work_book['AREA'] = [area_muni[0] for area_muni in area_munis]
-        output_work_book['MUNICIPALITY'] = [area_muni[1] for area_muni in area_munis]
+        # Update the DataFrame with predictions only for the rows that were used for prediction
+        output_work_book.loc[prediction_mask, 'AREA'] = [area_muni[0] for area_muni in area_munis]
+        output_work_book.loc[prediction_mask, 'MUNICIPALITY'] = [area_muni[1] for area_muni in area_munis]
 
         # Save the updated DataFrame to the same Excel file
         output_work_book.to_excel(output_file_path, index=False)
+
+        set_progress((total_files + 4) / work_progress * 100)
+
+        func.highlight_n_check_prediction(output_file_path)
+        func.auto_fit_columns(output_file_path)
 
         set_progress((total_files + 5) / work_progress * 100)
 
@@ -198,7 +223,10 @@ def merge():
 
     except Exception as e:
         message = f"{e}"
-        os.remove(output_file_path)
+        # Check if output_file_path is not None before trying to remove it
+        if output_file_path is not None:
+            os.remove(output_file_path)
+        
 
     data_to_return = {'message': message, 'file_path': folder_path, 'status': status}
 
@@ -208,38 +236,41 @@ def merge():
 
 @app.route('/feed', methods=['POST'])
 def feed():
-    
     status = False
     
-    file = request.files['file']
-    file_name = secure_filename(file.filename)
-    file_path = os.path.join(app.config['UPLOADED_FILES_DEST'], file_name)
-    file.save(file_path)
+    file = request.files.get('file')  # Safely get file
+    if file:
+        try:
+            file_name = secure_filename(file.filename)
+            file_path = os.path.join(app.config['UPLOADED_FILES_DEST'], file_name)
+            file.save(file_path)
 
-    df1 = pd.read_excel(file_path)
+            df1 = pd.read_excel(file_path)
 
-    model_path = os.path.join(app.config['SOURCE_FILES_DEST'], "model.xlsx")
-    df2 = pd.read_excel(model_path)
+            model_path = os.path.join(app.config['SOURCE_FILES_DEST'], "model.xlsx")
+            df2 = pd.read_excel(model_path)
 
-    try:        
-        if list(df1.columns) == list(df2.columns):
-            combine_df = pd.concat([df2, df1], ignore_index=True)
-        
-            combine_df.to_excel(model_path, index=False)
+            if list(df1.columns) == list(df2.columns):
+                combined_df = pd.concat([df2, df1], ignore_index=True)
+                combined_df.to_excel(model_path, index=False)
 
-            train_model_save_joblib()
-            status = True
-            
-            return jsonify({"message": "New data has been fed", "status": status})
-        else:
-            return jsonify({"message": "Wrong column format", "status": status})
-    except:
-        df2.to_excel(model_path, index=False)
-    finally:
-        os.remove(file_path)  # Clean up the uploaded file
+                train_model_save_joblib()
 
-
-    app.run()
+                status = True
+                return jsonify({"message": "New data has been fed", "status": status})
+            else:
+                return jsonify({"message": "Wrong column format", "status": status})
+        except FileNotFoundError:
+            return jsonify({"message": "File not found", "status": status}), 404
+        except pd.errors.ParserError:
+            return jsonify({"message": "Error parsing Excel file", "status": status}), 400
+        except Exception as e:
+            return jsonify({"message": f"An error occurred: {str(e)}", "status": status}), 500
+        finally:
+            if os.path.exists(file_path):
+                os.remove(file_path)  # Clean up the uploaded file
+    else:
+        return jsonify({"message": "No file uploaded", "status": status}), 400
 
 @app.route('/', methods=['GET'])
 def index():
